Update
Many bugs have been removed

+++ src/commands/custom-commands/delete.js
@@ -3,13 +3,23 @@ const Schema = require("../../database/models/customCommandAdvanced");
 
 module.exports = async (client, interaction, args) => {
     const cmdname = interaction.options.getString('command');
+
+    // Проверка наличия специальных символов
+    const regex = /[!@#$%^&*()_+\-=\[\]{};':"\\|,.<>\/?]+/;
+    if (regex.test(cmdname)) {
+        return client.errNormal({
+            error: "Command name contains special characters, please use only alphanumeric characters!",
+            type: 'editreply'
+        }, interaction);
+    }
+
     Schema.findOne({ Guild: interaction.guild.id, Name: cmdname.toLowerCase() }, async (err, data) => {
-        console.log(data)
+        console.log(data);
         if (data) {
             Schema.findOneAndDelete({ Guild: interaction.guild.id, Name: cmdname.toLowerCase() }).then(async () => {
-                var commands = await interaction.guild.commands.fetch()
-                var command = await commands.find((cmd => cmd.name == cmdname.toLowerCase()))
-                if(!command) return client.errNormal({ error: "Unable to find this command!", type: 'editreply' }, interaction );
+                var commands = await interaction.guild.commands.fetch();
+                var command = await commands.find((cmd) => cmd.name == cmdname.toLowerCase());
+                if (!command) return client.errNormal({ error: "Unable to find this command!", type: 'editreply' }, interaction);
                 await interaction.guild.commands.delete(command.id);
 
                 client.succNormal({
@@ -21,13 +31,9 @@ module.exports = async (client, interaction, args) => {
                     }],
                     type: 'editreply'
                 }, interaction);
-            })
-        }
-        else {
+            });
+        } else {
             client.errNormal({ error: "Unable to find this command!", type: 'editreply' }, interaction);
         }
-    })
-
-}
-
- 
+    });
+};
